requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
email-validator
pytest
pytest-asyncio
httpx
asyncpg
bcrypt
Pillow
openpyxl
pandas
celery
redis
python-magic-bin
